+++ devtools/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v0.1.0] devtools
 ### Releases
 - [NPM v0.1.0] https://www.npmjs.com/package/@avalanche-sdk/devtools/v/0.1.0 - devtools
+
+## 2025-07-15 18:41:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.582.0 (2.657.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.0] devtools
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/@avalanche-sdk/devtools/v/0.2.0 - devtools